+++ final02/src/components/Home.module.css
@@ -964,3 +964,45 @@
   background: #66c0f4;
   color: #ffffff;
 }
+
+/* Home.module.css에 추가 또는 수정 */
+
+.featuredGame {
+  position: relative;
+  width: 100%;
+  height: 460px; /* BeatSaber 이미지와 같은 비율로 설정 */
+  overflow: hidden;
+  cursor: pointer;
+}
+
+.featuredGameImage {
+  width: 100%;
+  height: 100%;
+  object-fit: contain; /* cover에서 contain으로 변경 */
+  background-color: #1b2838; /* 어두운 배경색 추가 */
+}
+
+.featuredGameInfo {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  padding: 20px;
+  background: linear-gradient(to top, rgba(0,0,0,0.8) 0%, rgba(0,0,0,0) 100%);
+  color: white;
+}
+
+.featuredGameTitle {
+  font-size: 24px;
+  margin-bottom: 10px;
+}
+
+.tag {
+  display: inline-block;
+  padding: 4px 8px;
+  margin: 0 4px 4px 0;
+  background: rgba(103, 193, 245, 0.2);
+  border-radius: 3px;
+  font-size: 12px;
+  color: #67c1f5;
+}
