+++ final02/src/components/library/Library.js
@@ -17,9 +17,9 @@ const Library = () => {
   // 공통 이미지 로딩 함수
   const loadGameImage = async (gameNo) => {
     try {
-      const response = await axios.get(`http://localhost:8080/game/image/${gameNo}`);
+      const response = await axios.get(`/game/image/${gameNo}`);
       if (response.data && response.data.length > 0) {
-        return `http://localhost:8080/game/download/${response.data[0].attachmentNo}`;
+        return `${process.env.REACT_APP_BASE_URL}/game/download/${response.data[0].attachmentNo}`;
       }
     } catch (error) {
       console.error(`Error loading image for game ${gameNo}:`, error);
@@ -55,7 +55,7 @@ const Library = () => {
   // 추천 게임 리스트 로드
   const loadGameList = useCallback(async () => {
     try {
-      const response = await axios.get("http://localhost:8080/game/");
+      const response = await axios.get("/game/");
       const shuffledGames = response.data.sort(() => Math.random() - 0.5).slice(0, 6);
 
       const gameListWithImages = await Promise.all(
